Add integration tests and CI Fix wrong nstream in plain_stream
